refactor: move monitoring and rules defaulters and validators to types
