+++ pkg/operator/rules.go
@@ -19,23 +19,18 @@ import (
 	"fmt"
 
 	"github.com/go-logr/logr"
-	yaml "gopkg.in/yaml.v3"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
-	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/export"
 	monitoringv1 "github.com/GoogleCloudPlatform/prometheus-engine/pkg/operator/apis/monitoring/v1"
-	"github.com/GoogleCloudPlatform/prometheus-engine/pkg/rules"
 )
 
 const (
@@ -240,7 +235,7 @@ func (r *rulesReconciler) ensureRuleConfigs(ctx context.Context, projectID, loca
 		return fmt.Errorf("list rules: %w", err)
 	}
 	for _, rs := range rulesList.Items {
-		result, err := generateRules(&rs, projectID, location, cluster)
+		result, err := rs.RuleGroupsConfig(projectID, location, cluster)
 		if err != nil {
 			// TODO(freinartz): update resource condition.
 			logger.Error(err, "converting rules failed", "rules_namespace", rs.Namespace, "rules_name", rs.Name)
@@ -256,7 +251,7 @@ func (r *rulesReconciler) ensureRuleConfigs(ctx context.Context, projectID, loca
 		return fmt.Errorf("list cluster rules: %w", err)
 	}
 	for _, rs := range clusterRulesList.Items {
-		result, err := generateClusterRules(&rs, projectID, location, cluster)
+		result, err := rs.RuleGroupsConfig(projectID, location, cluster)
 		if err != nil {
 			// TODO(freinartz): update resource condition.
 			logger.Error(err, "converting rules failed", "clusterrules_name", rs.Name)
@@ -272,7 +267,7 @@ func (r *rulesReconciler) ensureRuleConfigs(ctx context.Context, projectID, loca
 		return fmt.Errorf("list global rules: %w", err)
 	}
 	for _, rs := range globalRulesList.Items {
-		result, err := generateGlobalRules(&rs)
+		result, err := rs.RuleGroupsConfig()
 		if err != nil {
 			// TODO(freinartz): update resource condition.
 			logger.Error(err, "converting rules failed", "globalrules_name", rs.Name)
@@ -292,107 +287,4 @@ func (r *rulesReconciler) ensureRuleConfigs(ctx context.Context, projectID, loca
 		return fmt.Errorf("update generated rules: %w", err)
 	}
 	return nil
-}
-
-func generateRules(apiRules *monitoringv1.Rules, projectID, location, cluster string) (string, error) {
-	rs, err := rules.FromAPIRules(apiRules.Spec.Groups)
-	if err != nil {
-		return "", fmt.Errorf("converting rules failed: %w", err)
-	}
-	if err := rules.Scope(&rs, map[string]string{
-		export.KeyProjectID: projectID,
-		export.KeyLocation:  location,
-		export.KeyCluster:   cluster,
-		export.KeyNamespace: apiRules.Namespace,
-	}); err != nil {
-		return "", fmt.Errorf("isolating rules failed: %w", err)
-	}
-	result, err := yaml.Marshal(rs)
-	if err != nil {
-		return "", fmt.Errorf("marshalling rules failed: %w", err)
-	}
-	return string(result), nil
-}
-
-func generateClusterRules(apiRules *monitoringv1.ClusterRules, projectID, location, cluster string) (string, error) {
-	rs, err := rules.FromAPIRules(apiRules.Spec.Groups)
-	if err != nil {
-		return "", fmt.Errorf("converting rules failed: %w", err)
-	}
-	if err := rules.Scope(&rs, map[string]string{
-		export.KeyProjectID: projectID,
-		export.KeyLocation:  location,
-		export.KeyCluster:   cluster,
-	}); err != nil {
-		return "", fmt.Errorf("isolating rules failed: %w", err)
-	}
-	result, err := yaml.Marshal(rs)
-	if err != nil {
-		return "", fmt.Errorf("marshalling rules failed: %w", err)
-	}
-	return string(result), nil
-}
-
-func generateGlobalRules(apiRules *monitoringv1.GlobalRules) (string, error) {
-	rs, err := rules.FromAPIRules(apiRules.Spec.Groups)
-	if err != nil {
-		return "", fmt.Errorf("converting rules failed: %w", err)
-	}
-	if err := rules.Scope(&rs, map[string]string{}); err != nil {
-		return "", fmt.Errorf("isolating rules failed: %w", err)
-	}
-	result, err := yaml.Marshal(rs)
-	if err != nil {
-		return "", fmt.Errorf("marshalling rules failed: %w", err)
-	}
-	return string(result), nil
-}
-
-type rulesValidator struct {
-	opts Options
-}
-
-func (v *rulesValidator) ValidateCreate(_ context.Context, o runtime.Object) (admission.Warnings, error) {
-	_, err := generateRules(o.(*monitoringv1.Rules), "test_project", "test_location", "test_cluster")
-	return nil, err
-}
-
-func (v *rulesValidator) ValidateUpdate(ctx context.Context, _, o runtime.Object) (admission.Warnings, error) {
-	return v.ValidateCreate(ctx, o)
-}
-
-func (v *rulesValidator) ValidateDelete(_ context.Context, _ runtime.Object) (admission.Warnings, error) {
-	return nil, nil
-}
-
-type clusterRulesValidator struct {
-	opts Options
-}
-
-func (v *clusterRulesValidator) ValidateCreate(_ context.Context, o runtime.Object) (admission.Warnings, error) {
-	_, err := generateClusterRules(o.(*monitoringv1.ClusterRules), "test_project", "test_location", "test_cluster")
-	return nil, err
-}
-
-func (v *clusterRulesValidator) ValidateUpdate(ctx context.Context, _, o runtime.Object) (admission.Warnings, error) {
-	return v.ValidateCreate(ctx, o)
-}
-
-func (v *clusterRulesValidator) ValidateDelete(_ context.Context, _ runtime.Object) (admission.Warnings, error) {
-	return nil, nil
-}
-
-type globalRulesValidator struct{}
-
-func (v *globalRulesValidator) ValidateCreate(_ context.Context, o runtime.Object) (admission.Warnings, error) {
-	_, err := generateGlobalRules(o.(*monitoringv1.GlobalRules))
-	return nil, err
-}
-
-func (v *globalRulesValidator) ValidateUpdate(ctx context.Context, _, o runtime.Object) (admission.Warnings, error) {
-	return v.ValidateCreate(ctx, o)
-}
-
-func (v *globalRulesValidator) ValidateDelete(_ context.Context, _ runtime.Object) (admission.Warnings, error) {
-	return nil, nil
 }
